pyproject.toml
# SPDX-License-Identifier: Unlicense

[project]
name = "mcp-stdio-bridge"
version = "1.0.0"
description = "A generic MCP SSE <-> Stdio bridge with advanced security and SSL/TLS support."
license = { text = "Unlicense" }
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: The Unlicense (Unlicense)",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Communications :: File Sharing",
]
dependencies = [
    "mcp>=1.0.0",
    "starlette>=0.30.0",
    "uvicorn>=0.23.0",
    "anyio>=4.0.0",
    "sse-starlette>=1.6.0",
    "PyYAML>=6.0.0",
    "emoji>=2.0.0"
]
readme = "README.md"
requires-python = ">=3.10"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["mcp_stdio_bridge"]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "anyio>=4.0.0",
    "pytest-cov>=4.0.0",
    "pytest-timeout>=2.0.0",
    "httpx>=0.27.0",
    "jsonschema>=4.0.0",
]
dev = [
    "ruff>=0.3.0",
    "mypy>=1.9.0",
    "bandit>=1.7.0",
    "types-PyYAML",
]

[project.urls]
Homepage = "https://github.com/hackagadget/mcp-stdio-bridge"
Repository = "https://github.com/hackagadget/mcp-stdio-bridge"

[project.scripts]
mcp-stdio-bridge = "mcp_stdio_bridge.main:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=src --cov-report=term-missing"

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "**/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "pass",
    "raise ImportError",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "B", "S"]
ignore = ["S104"] # Ignore bind to all interfaces for this bridge

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101"]

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true
# This prevents the CLI and IDE from fighting over "extra" ignore codes
warn_unused_ignores = false
# This ensures both tools report the same error codes
show_error_codes = true

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
